fix: rename modelling for us english
