+++ dspy/modeling/backends/json.py
@@ -1,8 +1,8 @@
 import typing as t
 
-from dspy.modelling.lm import BaseLM
-from dspy.modelling.templates.base import BaseTemplate
-from dspy.modelling.templates.json import JSONTemplate
+from dspy.modeling.lm import BaseLM
+from dspy.modeling.templates.base import BaseTemplate
+from dspy.modeling.templates.json import JSONTemplate
 from dspy.primitives.example import Example
 from dspy.primitives.prediction import Completions
 from dspy.signatures.signature import Signature
