+++ dspy/modeling/templates/base.py
