+++ dspy/modeling/templates/json.py
