+++ dspy/modeling/templates/text.py
