+++ tests/predict/test_chain_of_thought_with_hint.py
@@ -1,7 +1,7 @@
 import dspy
 from dspy import ChainOfThoughtWithHint
 from dspy.utils import DummyLanguageModel, DummyLM
-from dspy.modelling import TemplateBackend
+from dspy.modeling import TemplateBackend
 
 
 def test_cot_with_no_hint():
